fix: linspace dtype error

+++ networks/scheduler/ddpm.py
@@ -6,7 +6,7 @@ class DDPMSampler:
     def __init__(self, generator: torch.Generator, num_training_steps=1000, beta_start: float = 0.00085, beta_end: float = 0.0120):
         # Params "beta_start" and "beta_end" taken from: https://github.com/CompVis/stable-diffusion/blob/21f890f9da3cfbeaba8e2ac3c425ee9e998d5229/configs/stable-diffusion/v1-inference.yaml#L5C8-L5C8
         # For the naming conventions, refer to the DDPM paper (https://arxiv.org/pdf/2006.11239.pdf)
-        self.betas = torch.linspace(beta_start ** 0.5, beta_end ** 0.5, num_training_steps, dtype=torch.float16) ** 2
+        self.betas = torch.linspace(beta_start ** 0.5, beta_end ** 0.5, num_training_steps) ** 2
         self.alphas = 1.0 - self.betas
 
         self.alphas_cumprod = torch.cumprod(self.alphas.to(torch.float32), dim=0).to(torch.float16)
